feat: Redis DI 및 로그인, 토큰 재발급에 redis 추가

+++ api/src/main/java/kr/bora/api/common/util/RedisUtil.java
@@ -16,7 +16,7 @@ public class RedisUtil {
     private final RedisTemplate<String, Object> redisBlackListTemplate;
 
 
-    public void set(String key, Object o, int minutes) {
+    public void set(String key, Object o, Long minutes) {
         redisTemplate.setValueSerializer(new Jackson2JsonRedisSerializer(o.getClass()));
         redisTemplate.opsForValue().set(key, o, minutes, TimeUnit.MINUTES);
     }
